move test classes to commonTest
